[CP-2341] Implement logic to handle select device
